Iteration 1
There are two changes that I have made so far.
 - I created a color palette in kuler with the dark blue and the light blue
from the logo as one corner of my triad color scheme. I set the backgrounds
of the existing classes to colors from this scheme. I set the background of...
	* the body to dark blue,
	* the container to light yellow,
	* the header to light blue,
	* the article headings to dark peach(ish?)
I will set the navigation pannel to light peach(ish) after I create it.

 - I made a text file will notes about how everything should be changed.